Client/goalforge/Server/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-restful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
